Update where svg serialization occurs
