+++ packages/core/util/declare.d.ts
